% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bootstrap.config.matrices}
\alias{bootstrap.config.matrices}
\title{Performing a single bootstrap}
\usage{
bootstrap.config.matrices(config_matrices)
}
\arguments{
\item{config_matrices}{The result from compute.config.matrices().}
}
\value{
An n x n matrix of RV coefficients for the bootstrapped data, where n is the number of datasets.
}
\description{
Helper function for run.bootstraps(). It's unlikely you'll ever need to run this function directly.
}
