% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{inner.product}
\alias{inner.product}
\title{Inner product similarity.}
\usage{
inner.product(x, y)
}
\arguments{
\item{x}{A vector of numbers.}

\item{y}{A vector of numbers.}
}
\value{
The inner product similarity between x and y.
}
\description{
Computes the inner product between x and y.
}
\examples{
set.seed(2)
n = 100
x = rnorm(n)
y = rnorm(n)
inner.product(x, y)
}
