% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{jaccard}
\alias{jaccard}
\title{Jaccard similarity.}
\usage{
jaccard(x, y)
}
\arguments{
\item{x}{A vector of zeroes and ones.}

\item{y}{A vector of zeroes and ones.}
}
\value{
The Jaccard similarity between x and y.
}
\description{
Computes the Jaccard similarity between x and y. When both x and
y only contain zeroes, the Jaccard similarity it not defined. This
function returns zero for that specific case.
}
\examples{
set.seed(2)
n = 100
x = rbinom(n, 1, 0.5)
y = rbinom(n, 1, 0.5)
jaccard(x, y)
}
